Add files via upload
These are the base files for the game, which was part of an assignment for my fundamentals of computer science 2 class in 2019. I went a little beyond the scope of the assignment and scored some extra credit points. They do however, depend on some code designed for the class. I think I'm allowed to publish that code, I'm not sure, but I also need to find it. As of now, it is not in this repository, so I don't think this code should run. I'll look into finding it to publish.